+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-> NOTE: LiveStore is still in alpha and releases can include breaking changes. See [state of the project](https://preview.livestore.dev/docs/reference/state-of-the-project/) for more info.
+> NOTE: LiveStore is still in beta and releases can include breaking changes. See [state of the project](https://docs.livestore.dev/evaluating/state-of-the-project/) for more info.
 > LiveStore is following a semver-like release strategy where breaking changes are released in minor versions before the 1.0 release.
 
 ## 0.3.0